+++ src/mappings.ts
@@ -11,6 +11,7 @@ export const HUBSPOT_DEFAULT_FIELD_MAPPINGS: Record<
   accounts: {
     name: "name",
     domain: "domain",
+    parentAccountId: "hs_parent_company_id",
     industry: "industry",
     employeeCount: "numberofemployees",
     annualRevenue: "annualrevenue",
@@ -55,6 +56,7 @@ export const SALESFORCE_DEFAULT_FIELD_MAPPINGS: Record<
     id: "Id",
     name: "Name",
     domain: "Website",
+    parentAccountId: "ParentId",
     industry: "Industry",
     employeeCount: "NumberOfEmployees",
     annualRevenue: "AnnualRevenue",